Introduce a PullRequestParser object.
I want to decouple the javascript from directly modifying things
globally.  I want the global state to be passed in (for example the DOM
elements that are expected to be used, the options for the library,
etc.).

To do so I am introducing a PullRequestParser prototype that is
constructed from the actual app.  Eventually, this should provide a way
to customize things to a greater detail.