Allow user to enter organization name instead of repositories.
Rather than requiring the user to enter every single repository for an
organization (and continue to do so to keep it up-to-date), allow just
the organization name to be entered.  All of the organizations
repositories will be looked up and all of the pull requests for each of
them will be shown.