Add pagination handling.
In order to ensure all items are being looked at (especially for things
like an organization's repositories which could be numerous), add
automatic pagination handling to the github api client in here.

Github does its pagination via the "Link" header.  According to their
documentation, the only reliable way to do the pagination is by
following the headers.  While many of their resources could be guessed
(page=1, page=2, page=3, up to the page mentioned in the "last" link)
this can't be relied upon and some of their api endpoints return next
links based off of hashes, not pages.  This means that we can't try to
optimize and request multiple pages at once, we can only paginate based
off the links they give us and do so recursively one at a time.

We do have to interject ourselves into the jquery deferred chain rather
than sticking to bluebird promises because we need access to the xhr
object to get the http headers.  Thankfully things are simple here and
things read nicely.

The link header parsing is a bit funky, but without pulling in a
dependency on some third party library, it's not too bad and should be
reliable for github's api at least.